[hotfix][tests] Extract class InfiniteIntegerSource
